kafka-sync/requirements.txt
# Kafka
kafka-python==2.0.2

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
